// src/assets/lenguajes.d.ts
export type Lenguajes = {
    [key: string]:{
        "headerBar":{
            "nav":{
                "inicio": string
                "historia": string
                "proposito": string
                "comunidad": string
            },
            "button": string
        },
        "main":{
            "section1":{
                "container0": {
                    "title": string
                    "text": string
                },
                "container1":{
                    "title": string
                    "text": string
                },
                "container2":{
                    "title": string
                    "text": string
                },
                "container3":{
                    "title": string
                    "text": string
                },
                "container4":{
                    "title": string
                    "c4t1": string
                    "c4t2": string
                    "c4t3": string
                    "c4t4": string
                    "c4t5": string
                    "c4t6":string
                } 
            },
            "section2":{
                "title": string
                "text": string
                "cards":{
                    "card1": string
                    "card2": string
                    "card3": string
                },
                "text2": string
            },
            "section3":{
                "title": string
                "text": string
                "cards":{
                    "card1":{
                        "title": string
                        "text": string
                    },
                    "card2":{
                        "title": string
                        "text": string
                    },
                    "card3":{
                        "title": string
                        "text": string
                    }
                }
            }
        },
        "footer":{
            "button": string
        }
    }
}